requirements.txt
fastapi==0.111.0
strawberry-graphql[fastapi]==0.211.0
uvicorn[standard]==0.30.1
